\name{get_dummy_bubbletree}



\alias{get_dummy_bubbletree}




\title{
Build bubbletree with external clustering solution
}



\description{
Often when inspecting scRNA-seq data we want to have a more in-depth look at
the cellular structure in one bubble or a set of bubbles of an existing
bubbletree without affecting the remaining bubbles. This can be achieved with
the function update_bubbletree, which performs additional clustering of cells
within existing bubbles, followed by updating of the hierarchical dendrogram.

**Important note**: this function should only be used for data exploration! Optimal
clustering solution is not guaranteed.}



\usage{
get_dummy_bubbletree(x,
                     cs,
                     B = 100,
                     N_eff = 100,
                     cores = 1,
                     seed = NA,
                     round_digits = 2,
                     show_simple_count = F)
}


\arguments{
\item{x}{numeric matrix (\eqn{A^{n\times f}} with \eqn{n} cells, and \eqn{f}
low-dimensional projections of the original single cell RNA-seq dataset)}
\item{cs}{vector, cluster assignments computed externally}
\item{N_eff}{integer, number of cells to draw randomly from each cluster when
computing inter-cluster distances. Maximum available number of cells are used
for clusters that contain lower number of cells than N_eff}
\item{cores}{integer, number of PC cores for parallel execution}
\item{seed}{number, for reproducing results}
\item{round_digits}{integer, number of decimal places to keep when showing the
relative frequency of cells in each bubble}
\item{show_simple_count}{logical, if show_simple_count=T, cell counts in each
bubble will be divided by 1,000 to improve readability. This is only useful for
samples that are composed of millions of cells.}
}


\details{
This function is nearly identical to get_bubbletree but does skips the k-means
clustering step. See documentation of get_bubbletree. The output is again a
bubbletree object
}


\value{
\item{A}{input x matrix}
\item{km}{k-means clustering results identical to those generated by function
          k-means from R-package stats}
\item{ph}{boot_ph: bootstrap dendrograms \eqn{H_b}; main_ph: consensus
          dendrogram \eqn{\hat{H}}}
\item{pair_dist}{inter-cluster distances used to generate the dendrograms}
\item{k}{number of clusters}
\item{cluster}{cluster assignments of each cell}
\item{input_par}{list of all input parameters}
\item{tree}{ggtree bubbletree object}
\item{tree_meta}{meta-data associated with the bubbletree}
}


\author{
  Simo Kitanovski <simo.kitanovski@uni-due.de>
}


\examples{
# input data

data(A)

head(A)

# dummy clusters
cs <- sample(x = LETTERS, size = nrow(A), replace = T)

db <- dummy_bubbletree(x = A,
                       cs,
                       B = 100,
                       N_eff = 100,
                       cores = 1)


}

\seealso{get_k, update_bubbletree, get_bubbletree, get_gini, get_gini_k, A}
