\docType{package}
\name{scBubbletree-package}
\alias{scBubbletree-package}
\alias{scBubbletree}

\title{
The R package scBubbletree
}

\description{
Method for quantitative visualization of single cell RNA-seq data
}

\details{
This package contains functions for clustering, hierarchical grouping of 
clusters and visualization of scRNA-seq data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/snaketron/scBubbletree}
  \item Report bugs at \url{https://github.com/snaketron/scBubbletree/issues}
}
}

\author{
  
Authors and maintainers:
\itemize{
  \item Simo Kitanovski \email{simokitanovski@gmail.com} 
  (\href{https://orcid.org/0000-0003-2909-5376}{ORCID})
}
}
