\name{update_bubbletree}



\alias{update_bubbletree}




\title{
In depth exploration of specific bubbles of a bubbletree by additional and
specific clustering
}



\description{
Often when inspecting scRNA-seq data we want to have a more in-depth look at
the cellular structure in one bubble or a set of bubbles of an existing
bubbletree without affecting the remaining bubbles. This can be achieved with
the function update_bubbletree, which performs additional clustering of cells
within existing bubbles, followed by updating of the hierarchical dendrogram.

**Important note**: this function should only be used for data exploration! Optimal
clustering solution is not guaranteed.}



\usage{
update_bubbletree(btd,
                  updated_bubbles,
                  k,
                  cores = 1)
}


\arguments{
\item{btd}{bubbletree object, input parameters specified in btd are reused}
\item{updated_bubbles}{vector, names of bubbles in btd to be updated}
\item{k}{integer, number of clusters (\eqn{k} should be larger than the number
of bubbles in vector updated_bubbles)}
\item{cores}{integer, number of PC cores for parallel execution}
}


\details{
This function performs nearly identical operation as get_bubbletree but
for a subset of the data. See documentation of get_bubbletree. The output
is again a bubbletree object}

\value{
\item{A}{input x matrix}
\item{km}{k-means clustering results identical to those generated by function
          k-means from R-package stats}
\item{ph}{boot_ph: bootstrap dendrograms \eqn{H_b}; main_ph: consensus
          dendrogram \eqn{\hat{H}}}
\item{pair_dist}{inter-cluster distances used to generate the dendrograms}
\item{k}{number of clusters}
\item{cluster}{cluster assignments of each cell}
\item{input_par}{list of all input parameters}
\item{tree}{ggtree bubbletree object}
\item{tree_meta}{meta-data associated with the bubbletree}
}


\author{
  Simo Kitanovski <simo.kitanovski@uni-due.de>
}


\examples{
# input data

data(A)

head(A)

b <- get_bubbletree(x = A,
                    k = 18,
                    N_eff = 100)


ub <- update_bubbletree(x = btd,
                        updated_bubbles = c("1", "2")
                        k = 5)

}

\seealso{get_k, get_bubbletree, get_dummy_bubbletree, get_gini, get_gini_k, A}
